server/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-migrate
python-dotenv
passlib[bcrypt]